pyproject.toml
[project]
name = "hw-day3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "chromadb>=0.5.0",
    "python-dotenv",
    "openai",
    "requests",
]